+++ modules/client/src/multisig/internal/utils.ts
@@ -8,7 +8,6 @@ import {
 import {
   MultisigProposal,
   MultisigProposalListItem,
-  SubgraphMultisigApproversListItem,
   SubgraphMultisigProposal,
   SubgraphMultisigProposalListItem,
 } from "../interfaces";
@@ -26,6 +25,11 @@ export function toMultisigProposal(
   const endDate = new Date(
     parseInt(proposal.endDate) * 1000,
   );
+  const executionDate = proposal.executionDate
+    ? new Date(
+      parseInt(proposal.executionDate) * 1000,
+    )
+    : null;
   return {
     id: proposal.id,
     dao: {
@@ -40,10 +44,17 @@ export function toMultisigProposal(
       resources: metadata.resources,
       media: metadata.media,
     },
+    settings: {
+      onlyListed: proposal.plugin.onlyListed,
+      minApprovals: parseInt(proposal.plugin.minApprovals),
+    },
+    creationBlockNumber: parseInt(proposal.creationBlockNumber) || 0,
     creationDate,
     startDate,
     endDate,
-    executionTxHash: proposal.executionTxHash || "",
+    executionDate,
+    executionBlockNumber: parseInt(proposal.executionBlockNumber) || null,
+    executionTxHash: proposal.executionTxHash || null,
     actions: proposal.actions.map(
       (action: SubgraphAction): DaoAction => {
         return {
@@ -58,7 +69,7 @@ export function toMultisigProposal(
     // Missing executable property from subgraph
     status: proposal.executed ? ProposalStatus.EXECUTED : ProposalStatus.ACTIVE,
     approvals: proposal.approvers.map(
-      (approver: SubgraphMultisigApproversListItem) => approver.approver.id,
+      (approver) => approver.id.slice(0, 42),
     ),
   };
 }
@@ -83,8 +94,14 @@ export function toMultisigProposalListItem(
       title: metadata.title,
       summary: metadata.summary,
     },
-    approvals: parseInt(proposal.approvals),
-    startDate, 
+    approvals: proposal.approvers.map(
+      (approver) => approver.id.slice(0, 42),
+    ),
+    settings: {
+      onlyListed: proposal.plugin.onlyListed,
+      minApprovals: parseInt(proposal.plugin.minApprovals),
+    },
+    startDate,
     endDate,
     // TODO
     // change to computeProposalStatus