+++ modules/client/src/tokenVoting/interfaces.ts
@@ -2,16 +2,16 @@
 import { BigNumber } from "@ethersproject/bignumber";
 import {
   ContractVotingSettings,
+  CreateMajorityVotingProposalParams,
   DaoAction,
   ExecuteProposalStepValue,
   GasFeeEstimation,
   ICanVoteParams,
   IClientCore,
-  CreateMajorityVotingProposalParams,
   IInterfaceParams,
   IProposalQueryParams,
-  MajorityVotingProposalSettings,
   IVoteProposalParams,
+  MajorityVotingProposalSettings,
   ProposalBase,
   ProposalCreationStepValue,
   ProposalListItemBase,
@@ -115,14 +115,15 @@ export type TokenVotingProposal = ProposalBase & {
   >;
   totalVotingWeight: bigint;
   creationBlockNumber: number;
-  executionDate: Date;
-  executionBlockNumber: number;
-  executionTxHash: string;
+  executionDate: Date | null;
+  executionBlockNumber: number | null;
+  executionTxHash: string | null;
 };
 
 export type TokenVotingProposalListItem = ProposalListItemBase & {
   token: Erc20TokenDetails | Erc721TokenDetails | null;
   result: TokenVotingProposalResult;
+  totalVotingWeight: bigint;
 };
 
 export type TokenVotingProposalResult = {
@@ -153,6 +154,7 @@ export type SubgraphTokenVotingProposalListItem = SubgraphProposalBase & {
   plugin: {
     token: SubgraphErc20Token | SubgraphErc721Token;
   };
+  totalVotingPower: string;
 };
 
 type SubgraphBaseToken = {