+++ modules/client/CHANGELOG.md
@@ -22,7 +22,9 @@ TEMPLATE:
 - Encoders and decoders for `upgradeToAndCallAction`
 - Examples for new decoders and encoders
 - JsDoc for new encoders and decoders
-
+- Add `settings` and `appovals` fields to `MultisigProposalBase`
+- Add `creationBlockNumber` `executionDate` and `creationBlockNumber` to `MultisigProposal`
+- Add `totalVotingWeight` to `tokenVotingListItem`
 ### Changed
 - `ICreateProposalParams` => `CreateMajorityVotingProposalParams`
 - `IProposalSettings` => `MajorityVotingProposalSettings`